+++ central/nodecomponentcveedge/datastore/datastore_sac_test.go
@@ -1,4 +1,4 @@
-package datastoretest
+package datastore
 
 import (
 	"context"
@@ -7,8 +7,8 @@ import (
 	componentCVEEdgeDataStore "github.com/stackrox/rox/central/componentcveedge/datastore"
 	"github.com/stackrox/rox/central/cve/converter/utils"
 	dackboxTestUtils "github.com/stackrox/rox/central/dackbox/testutils"
-	"github.com/stackrox/rox/central/nodecomponentcveedge/datastore"
 	"github.com/stackrox/rox/central/role/resources"
+	v1 "github.com/stackrox/rox/generated/api/v1"
 	"github.com/stackrox/rox/generated/storage"
 	"github.com/stackrox/rox/pkg/dackbox/edges"
 	"github.com/stackrox/rox/pkg/features"
@@ -16,6 +16,7 @@ import (
 	sacTestUtils "github.com/stackrox/rox/pkg/sac/testutils"
 	"github.com/stackrox/rox/pkg/scancomponent"
 	"github.com/stackrox/rox/pkg/search"
+	searchPkg "github.com/stackrox/rox/pkg/search"
 	"github.com/stackrox/rox/pkg/search/postgres"
 	"github.com/stretchr/testify/suite"
 )
@@ -35,7 +36,7 @@ type nodeComponentCVEEdgeDatastoreSACTestSuite struct {
 	suite.Suite
 
 	dackboxTestStore dackboxTestUtils.DackboxTestDataStore
-	datastore        datastore.DataStore
+	datastore        DataStore
 
 	testContexts map[string]context.Context
 }
@@ -46,7 +47,7 @@ func (s *nodeComponentCVEEdgeDatastoreSACTestSuite) SetupSuite() {
 	s.Require().NoError(err)
 	if features.PostgresDatastore.Enabled() {
 		pool := s.dackboxTestStore.GetPostgresPool()
-		s.datastore, err = datastore.GetTestPostgresDataStore(s.T(), pool)
+		s.datastore, err = GetTestPostgresDataStore(s.T(), pool)
 		s.Require().NoError(err)
 	} else {
 		rocksengine := s.dackboxTestStore.GetRocksEngine()
@@ -409,4 +410,56 @@ func (s *nodeComponentCVEEdgeDatastoreSACTestSuite) TestSearchRawEdges() {
 			}
 		})
 	}
+}
+
+// Helper type to have identical test code for both rocksdb/dackbox and postgres objects
+
+type nodeComponentCVEEdgeFromGenericStore struct {
+	genericStore componentCVEEdgeDataStore.DataStore
+}
+
+func imageComponentCVEEdgeToNodeComponentCVEEdge(edge *storage.ComponentCVEEdge) *storage.NodeComponentCVEEdge {
+	return &storage.NodeComponentCVEEdge{
+		Id:              edge.GetId(),
+		IsFixable:       edge.GetIsFixable(),
+		HasFixedBy:      &storage.NodeComponentCVEEdge_FixedBy{FixedBy: edge.GetFixedBy()},
+		NodeComponentId: edge.GetImageComponentId(),
+		NodeCveId:       edge.GetImageCveId(),
+	}
+}
+
+func (s nodeComponentCVEEdgeFromGenericStore) Count(ctx context.Context, q *v1.Query) (int, error) {
+	return s.genericStore.Count(ctx, q)
+}
+
+func (s nodeComponentCVEEdgeFromGenericStore) Search(ctx context.Context, q *v1.Query) ([]searchPkg.Result, error) {
+	return s.genericStore.Search(ctx, q)
+}
+
+func (s nodeComponentCVEEdgeFromGenericStore) SearchEdges(ctx context.Context, q *v1.Query) ([]*v1.SearchResult, error) {
+	return s.genericStore.SearchEdges(ctx, q)
+}
+
+func (s nodeComponentCVEEdgeFromGenericStore) SearchRawEdges(ctx context.Context, q *v1.Query) ([]*storage.NodeComponentCVEEdge, error) {
+	genericEdges, err := s.genericStore.SearchRawEdges(ctx, q)
+	if err != nil {
+		return nil, err
+	}
+	edges := make([]*storage.NodeComponentCVEEdge, 0, len(genericEdges))
+	for _, e := range genericEdges {
+		edges = append(edges, imageComponentCVEEdgeToNodeComponentCVEEdge(e))
+	}
+	return edges, nil
+}
+
+func (s nodeComponentCVEEdgeFromGenericStore) Exists(ctx context.Context, id string) (bool, error) {
+	return s.genericStore.Exists(ctx, id)
+}
+
+func (s nodeComponentCVEEdgeFromGenericStore) Get(ctx context.Context, id string) (*storage.NodeComponentCVEEdge, bool, error) {
+	genericEdge, found, err := s.genericStore.Get(ctx, id)
+	if err != nil || !found {
+		return nil, found, err
+	}
+	return imageComponentCVEEdgeToNodeComponentCVEEdge(genericEdge), true, nil
 }
